+++ CHANGELOG.md
@@ -6,9 +6,9 @@ If you like this project and find it useful, please consider giving it a star on
 
 ## [1.3.10] - 2024-07-05
 
-### Changed
+### Added
 
-- [dependencies]: Update dependencies.
+- [fabric]: Added sanitized fabricInfo to matterbridge in bridge mode and to the plugins in childbridge mode.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="./yellow-button.png" alt="Buy me a coffee" width="120">